centralizing the logging functions
