+++ src/laser_polio/__init__.py
@@ -3,6 +3,7 @@ from .pars import *  # noqa F403
 from .plotting import *  # noqa F403
 from .model import *  # noqa F403
 from .run_sim import *  # noqa F403
+from .logger import *  # noqa F403
 from pathlib import Path
 
 # from .seir_mpm import *
